Allow optional custom salt
Because I'm using this package in conjunction with the JS version, on the client side, I need to have matching salts but don't wish to expose my app's secret key to the world. This will now look for a package config specific salt before falling back to the app key.

Also mentioned in #21.